+++ src/components/layouts/Blogs.vue
@@ -1,0 +1,40 @@
+<template>
+  <div id="journal" class="text-left paddsection">
+    <b-container class="p-0">
+      <div class="section-title text-center">
+        <h2>Blogs</h2>
+      </div>
+    </b-container>
+
+    <b-container>
+      <div class="journal-block">
+        <b-row>
+          <b-col lg="4" md="6">
+            <div class="journal-info">
+              <a href="https://blog.francium.tech/learn-grpc-and-protobuf-with-golang-8456a2e64977?source=friends_link&sk=cfc3716536dd5972e6e5d744703194ea" target="_blank">
+                <img src="https://cdn-images-1.medium.com/max/800/1*JzeYG3sdWeKfcQ_A69m8oA.jpeg" class="img-responsive" alt="img">
+              </a>
+              <div class="journal-txt">
+                <h4>
+                  <a href="https://blog.francium.tech/learn-grpc-and-protobuf-with-golang-8456a2e64977?source=friends_link&sk=cfc3716536dd5972e6e5d744703194ea" target="_blank">
+                    Learn gRPC and Protobuf with GoLang
+                  </a>
+                </h4>
+                <p class="separator">
+                  This blog introduces you to gRPC and protobuf (protocol
+                  buffers) using GoLang.
+                </p>
+              </div>
+            </div>
+          </b-col>
+        </b-row>
+      </div>
+    </b-container>
+  </div>
+</template>
+
+<script>
+export default {
+  name: 'Blogs'
+}
+</script>
